#1334 cleanup timefield
