+++ lib/elementWrapper/elementWrapperFactory.js
@@ -30,10 +30,10 @@ const createElementWrapper = (elementType, query, attrValuePairs, _options, ...a
       return new ColorWrapper(selector, options, description);
     case 'RangeType':
       return new RangeWrapper(selector, options, description);
-    // case 'timeField':
-    //   return new TimeFieldWrapper(get, description, getIfExists);
-    // default:
-    //   return new ElementWrapper(get, description, getIfExists);
+    case 'TimeField':
+      return new TimeFieldWrapper(selector, options, description);
+    default:
+      return new ElementWrapper(selector, options, description);
   }
 };
