+++ lib/elementWrapper/timeFieldWrapper.js
@@ -1,7 +1,66 @@
 const TimeField = require('../elements/timeField');
 const ValueWrapper = require('./valueWrapper');
-const { firstElement } = require('./helper');
+const { firstElement, getElementGetter } = require('./helper');
+const { $function } = require('../elementSearch');
+
+function getTimeFieldElementWithLabel(label) {
+  const timeFields = [];
+  const inputTypes = ['date', 'datetime-local', 'month', 'time', 'week'];
+  function checkAndPushElement(elem) {
+    if (
+      elem.tagName &&
+      elem.tagName.toLowerCase() == 'input' &&
+      elem.type &&
+      inputTypes.includes(elem.type.toLowerCase())
+    ) {
+      timeFields.push(elem);
+    }
+  }
+  const matchingLabels = [...document.querySelectorAll('label')].filter((labelElem) => {
+    return labelElem.innerText.toLowerCase().includes(label.toLowerCase());
+  });
+  for (let matchingLabel of matchingLabels) {
+    const labelFor = matchingLabel.getAttribute('for');
+    if (labelFor) {
+      //check label with attribute for
+      const labelForElement = document.getElementById(labelFor);
+      checkAndPushElement(labelForElement);
+    } else {
+      // check child node of label tag
+      matchingLabel.childNodes.forEach((elem) => {
+        checkAndPushElement(elem);
+      });
+    }
+  }
+  //check label is inlined
+  const inputRadioElements = [
+    ...document.querySelectorAll(
+      'input[type="date" i],input[type="datetime-local" i],input[type="month" i],input[type="time" i],input[type="week" i]',
+    ),
+  ];
+  for (let inputRadioElement of inputRadioElements) {
+    if (
+      inputRadioElement.nextSibling.nodeType === Node.TEXT_NODE &&
+      inputRadioElement.nextSibling.wholeText.includes(label)
+    ) {
+      timeFields.push(inputRadioElement);
+    }
+  }
+  return timeFields;
+}
 class TimeFieldWrapper extends ValueWrapper {
+  _get = getElementGetter(
+    this.selector,
+    async () =>
+      await $function(
+        getTimeFieldElementWithLabel,
+        this.selector.label,
+        this.options.selectHiddenElements,
+      ),
+    'input[type="date" i],input[type="datetime-local" i],input[type="month" i],input[type="time" i],input[type="week" i]',
+    this.options.selectHiddenElements,
+  );
+
   async select(value) {
     const elem = await firstElement.apply(this);
     return await elem.select(value);
