+++ lib/taiko.js
@@ -2225,62 +2225,7 @@ module.exports.timeField = timeField;
 
 function timeField(attrValuePairs, _options = {}, ...args) {
   validate();
-  const { selector, options } = prepareParameters(attrValuePairs, _options, ...args);
-  function getTimeFieldElementWithLabel(label) {
-    const timeFields = [];
-    const inputTypes = ['date', 'datetime-local', 'month', 'time', 'week'];
-    function checkAndPushElement(elem) {
-      if (
-        elem.tagName &&
-        elem.tagName.toLowerCase() == 'input' &&
-        elem.type &&
-        inputTypes.includes(elem.type.toLowerCase())
-      ) {
-        timeFields.push(elem);
-      }
-    }
-    const matchingLabels = [...document.querySelectorAll('label')].filter((labelElem) => {
-      return labelElem.innerText.toLowerCase().includes(label.toLowerCase());
-    });
-    for (let matchingLabel of matchingLabels) {
-      const labelFor = matchingLabel.getAttribute('for');
-      if (labelFor) {
-        //check label with attribute for
-        const labelForElement = document.getElementById(labelFor);
-        checkAndPushElement(labelForElement);
-      } else {
-        // check child node of label tag
-        matchingLabel.childNodes.forEach((elem) => {
-          checkAndPushElement(elem);
-        });
-      }
-    }
-    //check label is inlined
-    const inputRadioElements = [
-      ...document.querySelectorAll(
-        'input[type="date" i],input[type="datetime-local" i],input[type="month" i],input[type="time" i],input[type="week" i]',
-      ),
-    ];
-    for (let inputRadioElement of inputRadioElements) {
-      if (
-        inputRadioElement.nextSibling.nodeType === Node.TEXT_NODE &&
-        inputRadioElement.nextSibling.wholeText.includes(label)
-      ) {
-        timeFields.push(inputRadioElement);
-      }
-    }
-    return timeFields;
-  }
-  const get = getElementGetter(
-    selector,
-    async () =>
-      await $function(getTimeFieldElementWithLabel, selector.label, options.selectHiddenElements),
-    'input[type="date" i],input[type="datetime-local" i],input[type="month" i],input[type="time" i],input[type="week" i]',
-    options.selectHiddenElements,
-  );
-
-  const description = desc(selector, 'label', 'Time field');
-  return createElementWrapper(get, description, 'timeField');
+  return createElementWrapper('TimeField', 'label', attrValuePairs, _options, ...args);
 }
 
 /**
